Move the room member table management out to a separate file
Trying to keep the size of matrix-room.c under control, move the membertable
management bits out to a separate file.

Also do away with the special handling for the first /sync of a room, and
instead *always* group together member arrivals/departures until after we've
finished handling the state table.