api: Allow binary data to be sent
Allow binary data to be sent after the headers.
This data may include a NUL byte and so everything gets
more complicated.
The data is sent raw on the wire.

It's used for uploading files.

We rename matrix_api_start to matrix_api_start_full with
the extra parameters and provide a matrix_api_start for
the callers who don't need the extra binary data.

Signed-off-by: Dr. David Alan Gilbert <dave@treblig.org>