event hooks
Some events need to do something else before the event can be
transmitted, and indeed the contents of the event aren't necessarily
known until that other thing has happened.

Add a hook that gets called when the event is ready to be sent
and when the event is to be free'd.

Signed-off-by: Dr. David Alan Gilbert <dave@treblig.org>