Create room (if join failed)
There is no hook in purple that allows to create a room
separately from joining a room.  In matrix there is a
distinction.  Therefore, try to join by either alias/id
and if this fails, create the room.

Currently this results in a disconnect but upon reconnecting
you become a member of the new room.

Signed-off-by: Damien Zammit <damien@zamaudio.com>