[matrix-room] Improve video thumbnail acquisition
When the thumbnail of a video is used, first check if a thumbnail_url is
available. If it is, directly get that, otherwise ask the server for a
thumbnail of the video.
When not using the thumbnail, after downloading the video warn the user
that it's somewhere.
Improve _[video|image]_download_* messages.