Remove _handle_incoming_image
Reduce duplication by handling all m.video, m.file, m.audio and m.image
in _handle_incoming_media. On incoming media, always print a chat
message with the download link and mimetype/size. If a thumbnail_url is
available and the thumbnail size is small, download and show that. Otherwise,
only for m_image, ask for a server generated thumbnail.