[matrix-room] Update thumbnail request to r0.3.0
With r0.3.0 thumbnail_info and thumbnail_url for m.image and m.file have
been moved from content to content.info. With this commit those objects
are looked for at the new location and if they're not found, the old
location is checked. Furthermore, if the thumbnail provided is too big,
we ask the server to make a thumbnail of it.